Move cann_device_test.rs to tests/test_cann_execution.rs
